+++ src/viewProviders/resources.ts
@@ -3,7 +3,6 @@ import { IconNames } from "../constants";
 import { getExtensionContext } from "../context";
 import { ccloudConnected, ccloudOrganizationChanged } from "../emitters";
 import { ExtensionContextNotSetError } from "../errors";
-import { CCloudEnvironmentGroup, getClustersByCCloudEnvironment } from "../graphql/environments";
 import { getLocalKafkaClusters } from "../graphql/local";
 import { getCurrentOrganization } from "../graphql/organizations";
 import { Logger } from "../logging";
@@ -35,13 +34,18 @@ type ResourceViewProviderData =
   | LocalKafkaCluster;
 
 export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceViewProviderData> {
+  // Did the user use the 'refresh' button / command to force a deep refresh of the tree?
+  private forceDeepRefresh: boolean = false;
+
   private _onDidChangeTreeData = new vscode.EventEmitter<
     ResourceViewProviderData | undefined | void
   >();
   readonly onDidChangeTreeData = this._onDidChangeTreeData.event;
 
-  refresh(): void {
+  refresh(forceDeepRefresh: boolean = false): void {
+    this.forceDeepRefresh = forceDeepRefresh;
     this._onDidChangeTreeData.fire();
+    this.forceDeepRefresh = false;
   }
 
   private treeView: vscode.TreeView<vscode.TreeItem>;
@@ -59,11 +63,14 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
 
     ccloudConnected.event((connected: boolean) => {
       logger.debug("ccloudConnected event fired", { connected });
+      // No need to force a deep refresh when the connection status changes because the
+      // preloader will have already begun loading resources due to also observing this event.
       this.refresh();
     });
 
     ccloudOrganizationChanged.event(() => {
-      this.refresh();
+      // Force a deep refresh of ccloud resources when the organization changes.
+      this.refresh(true);
     });
   }
 
@@ -96,12 +103,14 @@ export class ResourceViewProvider implements vscode.TreeDataProvider<ResourceVie
         // expand containers for kafka clusters, schema registry, flink compute pools, etc
         return element.children;
       } else if (element instanceof CCloudEnvironment) {
-        return await loadCCloudEnvironmentChildren(element);
+        return await getCCloudEnvironmentChildren(element);
       }
     } else {
       // --- ROOT-LEVEL ITEMS ---
       // NOTE: we end up here when the tree is first loaded
-      return await loadResources();
+      const forceDeepRefresh = this.forceDeepRefresh;
+      this.forceDeepRefresh = false;
+      return await loadResources(forceDeepRefresh);
     }
 
     return resourceItems;
@@ -113,7 +122,9 @@ export function getResourceViewProvider() {
   return ResourceViewProvider.getInstance();
 }
 
-async function loadResources(): Promise<ResourceViewProviderData[]> {
+async function loadResources(
+  forceDeepRefresh: boolean = false,
+): Promise<ResourceViewProviderData[]> {
   const resources: ResourceViewProviderData[] = [];
 
   // the section below will create a "Confluent Cloud" container item that will be either:
@@ -124,6 +135,10 @@ async function loadResources(): Promise<ResourceViewProviderData[]> {
   const preloader = CCloudResourcePreloader.getInstance();
 
   if (preloader.hasCCloudConnection()) {
+    if (forceDeepRefresh) {
+      // force a deep refresh of the resources next call to ensureResourcesLoaded()
+      preloader.reset();
+    }
     // Ensure all of the preloading is complete before referencing resource manager ccloud resources.
     await preloader.ensureResourcesLoaded();
 
@@ -181,22 +196,33 @@ async function loadResources(): Promise<ResourceViewProviderData[]> {
   return resources;
 }
 
-async function loadCCloudEnvironmentChildren(environment: CCloudEnvironment) {
+/**
+ * Return the children of a CCloud environmen (the Kafka clusters and Schema Registry).
+ * Called when expanding a CCloud environment tree item.
+ *
+ * Fetches from the cached resources in the resource manager.
+ *
+ * @param environment: The CCloud environment to get children for
+ * @returns
+ */
+async function getCCloudEnvironmentChildren(environment: CCloudEnvironment) {
   const subItems: (CCloudKafkaCluster | SchemaRegistryCluster)[] = [];
 
-  // load Kafka clusters and Schema Registry for the given environment, if they exist
-  const envGroup: CCloudEnvironmentGroup | null = await getClustersByCCloudEnvironment(environment);
-  if (!envGroup) {
-    // Return the empty array.
-    return subItems;
+  // Ensure all of the preloading is complete before referencing resource manager ccloud resources.
+  await CCloudResourcePreloader.getInstance().ensureResourcesLoaded();
+
+  const rm = getResourceManager();
+  // Get the Kafka clusters for this environment
+  const kafkaClusters = await rm.getCCloudKafkaClustersForEnvironment(environment.id);
+  subItems.push(...kafkaClusters);
+
+  // Schema registry?
+  const maybe_schema_registry: SchemaRegistryCluster | null =
+    await rm.getCCloudSchemaRegistryCluster(environment.id);
+  if (maybe_schema_registry) {
+    subItems.push(maybe_schema_registry);
   }
 
-  subItems.push(...envGroup.kafkaClusters);
-  if (envGroup.schemaRegistry) {
-    subItems.push(envGroup.schemaRegistry);
-  }
-
-  // TODO: add flink compute pools here
-
+  // TODO: add flink compute pools here ?
   return subItems;
 }